[FIX] - mi mytake

+++ src/main/java/com/javaproject/harang/entity/report/UserReports.java
@@ -29,7 +29,7 @@ public class UserReports {
 
     private Integer targetId;
 
-    private Integer score;
+    private double score;
 
     private LocalDate reportTime;
 }
